Added SDL to project for real time graphics and mouse/keyboard input.
This introduces lots of abstraction into this project, but it'll be  good practice with a (possibly) simpler API than DirectX11. At any rate, I don't think I'll have to worry about all the jargon and boilerplate code of graphics pipelines.

NOTE: the SDL file is NOT included in this repo - to get this working, follow the instructions at:

https://github.com/libsdl-org/SDL/blob/main/docs/INTRO-visualstudio.md